+++ ChangeLog.md
@@ -16,6 +16,8 @@ More detailed information about incremental changes can be found in the
   which was often the same.
 - Added IpoptApplication::Version() (C++ interface), GetIpoptVersion (C interface), and
   Ipopt::GetVersion() (Java interface) to retrieve version of Ipopt library [#824].
+- Fixed possible missing initialization of delta_x and delta_s in PDPerturbationHandler
+  in case ConsiderNewSystem failed [#834].
 
 ### 3.14.17 (2024-12-14)