+++ ChangeLog.md
@@ -22,6 +22,7 @@ More detailed information about incremental changes can be found in the
 - Added missing initialization of Filter Acceptor in case restoration phase is called
   when the fallback mechanism of BacktrackingLinearSearch has been activated [#834].
   If this happened in the first iteration, it lead to the use of uninitialized values.
+- Added missing return if symbolic factorization with MA57 (ma57ad, ma57as) failed [#834].
 
 ### 3.14.17 (2024-12-14)