+++ ChangeLog.md
@@ -7,6 +7,11 @@ More detailed information about incremental changes can be found in the
 
 ## 3.14
 
+### 3.14.20 (2025-xx-yy)
+
+- Fixed issue where Ipopt exceptions could not been caught from other libraries
+  on macOS with clang when Ipopt or the other library was build with `-fvisibility=hidden`.
+
 ### 3.14.19 (2025-07-30)
 
 - Fixed call to `getenv_s` on Windows, introduced with 3.14.17.